requirements.txt
cv2
pillow
flask